move recording code to GameContainer
