+++ packages/client/src/scenes/Game.tsx
@@ -10,22 +10,18 @@ import { View } from '../components';
 import { Card } from "rimble-ui";
 import GameManager from '../managers/GameManager';
 
-import { localSaveReplay, clientSaveTournamentReplay } from "../helpers/database";
-
 import GameResult from '../components/GameResult'
 
 import { toast } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
-
-declare global {
-  interface Window { MediaRecorder: any; }
-}
 
 interface IProps extends RouteComponentProps {
   roomId?: string;
   address?: string;
   drizzle?: any;
   drizzleState?: any;
+  startRecording: any;
+  stopRecording: any;
 }
 
 interface IState {
@@ -57,15 +53,6 @@ export default class Game extends Component<IProps, IState> {
   private client?: Client;
   public room?: Room;
 
-  private mediaRecorder: any;
-  private recordedBlobs: any;
-  private sourceBuffer: any;
-  private video: any;
-  private canvas: any;
-  private stream: any;
-  private mediaSource: any;
-  private gameOver: boolean;
-
   // BASE
   constructor(props: IProps) {
     super(props);
@@ -81,18 +68,6 @@ export default class Game extends Component<IProps, IState> {
   
   async componentDidMount() {
     await this.start();
-
-    this.mediaSource = new window.MediaSource();
-    // this.mediaSource.addEventListener('sourceopen', this.handleSourceOpen, false);
-  
-    this.canvas = document.querySelector('canvas');
-    this.video = document.querySelector('video');
-
-    if (this.canvas) {
-      this.stream = this.canvas.captureStream(); // frames per second
-      console.log('Started stream capture from canvas element: ', this.stream);
-    }
-
   }
 
   componentWillUnmount() {
@@ -277,13 +252,13 @@ export default class Game extends Component<IProps, IState> {
         {
           this.gameManager.hudLogAdd(`Game starts!`);
           this.gameManager.hudAnnounceAdd(`Game starts!`);
-          this.startRecording();
+          this.props.startRecording.call();
         }
         break;
       case 'stop':
         this.gameManager.hudLogAdd(`Game ends...`);
         this.setState({ gameOver: true });
-        await this.stopRecording();
+        await this.props.stopRecording.call();
         this.setState({
           showResult: true
         })
@@ -308,83 +283,6 @@ export default class Game extends Component<IProps, IState> {
         break;
     }
   }
-
-  // HANDLERS: Gameplay Recorder
-  startRecording() {
-    let options: any = { mimeType: 'video/webm' };
-    this.recordedBlobs = [];
-    try {
-        this.mediaRecorder = new window.MediaRecorder(this.stream, options);
-    } catch (e0) {
-        console.log('Unable to create MediaRecorder with options Object: ', e0);
-        try {
-            options = { mimeType: 'video/webm,codecs=vp9' };
-            this.mediaRecorder = new window.MediaRecorder(this.stream, options);
-        } catch (e1) {
-            console.log('Unable to create MediaRecorder with options Object: ', e1);
-            try {
-                options = 'video/vp8'; // Chrome 47
-                this.mediaRecorder = new window.MediaRecorder(this.stream, options);
-            } catch (e2) {
-                alert('MediaRecorder is not supported by this browser.\n\n' +
-                    'Try Firefox 29 or later, or Chrome 47 or later, ' +
-                    'with Enable experimental Web Platform features enabled from chrome://flags.');
-                console.error('Exception while creating MediaRecorder:', e2);
-                return;
-            }
-        }
-    }
-    console.log('Created MediaRecorder', this.mediaRecorder, 'with options', options);
-
-    this.mediaRecorder.onstop = (event) => {
-        // console.log('Recorder stopped: ', event);
-        // const superBuffer = new Blob(this.recordedBlobs, { type: 'video/webm' });
-        // this.video.src = window.URL.createObjectURL(superBuffer);
-    }
-
-    this.mediaRecorder.ondataavailable = (event) => {
-        if (event.data && event.data.size > 0) {
-            this.recordedBlobs.push(event.data);
-        }
-    }
-
-    this.mediaRecorder.start(100); // collect 100ms of data
-    // console.log('MediaRecorder started', this.mediaRecorder);
-  }
-
-  stopRecording = async () => {
-      this.mediaRecorder.stop();
-
-      // TODO: playerId = roomId? change to something more meaningful
-      const playerId = this.state.playerId;
-      const tournamentId = this.state.tournamentId || 'demo';
-      const time = this.gameManager.gameEndsAt - Date.now();
-
-      console.log('Recorded Blobs: ', this.recordedBlobs);
-
-      const replayDate = new Date();
-      const filename = "replay_" + playerId + "_" + replayDate.valueOf() + ".webm";
-      const options = {type:'video/webm'};
-
-      const file = new File(this.recordedBlobs, filename, options);
-  
-      // this.video.controls = true;
-
-
-      // TODO: move to web worker so it doesn't pause main thread
-      if (tournamentId === 'demo') {
-        localSaveReplay(playerId, tournamentId, time, file);
-      } else {
-        const recordFileHash = await clientSaveTournamentReplay(file);
-        this.setState({
-          recordFileHash
-        })
-        //const resultId = 1
-        //const result = await putTournamentResult(tournamentId, resultId, fileHash);
-        //console.log(result)
-      }
-  }
-
 
   // HANDLERS: GameManager
   handleActionSend = (action: any) => {
